added auth
added auth folder and reintroduced spring security

+++ src/main/java/trackour/trackour/models/User.java
@@ -40,7 +40,7 @@ public class User {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "uid")
-    private int uid;
+    private Long uid;
 
     @Column(name = "username")
     private String username;
